refactor: 브랜드 서비스 구조 리팩토링(#353)
